Flyttet validation ut i egen fil, fungerer fortsatt
