+++ apps/frontend/src/components/krav/Edit/KravEditPage.tsx
@@ -4,7 +4,6 @@ import { Form, Formik } from 'formik'
 import _ from 'lodash'
 import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
-import * as yup from 'yup'
 import { getEtterlevelserByKravNumberKravVersion } from '../../../api/EtterlevelseApi'
 import {
   TKravIdParams,
@@ -15,14 +14,7 @@ import {
   updateKrav,
 } from '../../../api/KravApi'
 import { GetKravData, IKravDataProps, TKravById } from '../../../api/KravEditApi'
-import {
-  EKravStatus,
-  EYupErrorMessage,
-  IKrav,
-  IKravId,
-  IKravVersjon,
-  TKravQL,
-} from '../../../constants'
+import { EKravStatus, IKrav, IKravId, IKravVersjon, TKravQL } from '../../../constants'
 import { TSection } from '../../../pages/EtterlevelseDokumentasjonPage'
 import { EListName, codelist } from '../../../services/Codelist'
 import { user } from '../../../services/User'
@@ -35,6 +27,7 @@ import { PageLayout } from '../../scaffold/Page'
 import { EditKravMultiOptionField } from './EditKravMultiOptionField'
 import { EditKravRelasjoner } from './EditKravRelasjoner'
 import { EditBegreper } from './KravBegreperEdit'
+import { kravSchema } from './KravEditPageValidation'
 import { KravSuksesskriterierEdit } from './KravSuksesskriterieEdit'
 import { KravVarslingsadresserEdit } from './KravVarslingsadresserEdit'
 import { RegelverkEdit } from './RegelverkEdit'
@@ -81,91 +74,6 @@ export const KravEditPage = () => {
 
   const [UtgaattKravMessage, setUtgaattKravMessage] = useState<boolean>(false)
   const [aktivKravMessage, setAktivKravMessage] = useState<boolean>(false)
-
-  const kravSchema = () =>
-    yup.object({
-      navn: yup.string().required('Du må oppgi et navn til kravet'),
-      suksesskriterier: yup.array().test({
-        name: 'suksesskriterierCheck',
-        message: EYupErrorMessage.PAAKREVD,
-        test: function (suksesskriterier) {
-          const { parent } = this
-          if (parent.status === EKravStatus.AKTIV) {
-            return suksesskriterier &&
-              suksesskriterier.length > 0 &&
-              suksesskriterier.every((s) => s.navn)
-              ? true
-              : false
-          }
-          return true
-        },
-      }),
-      hensikt: yup.string().test({
-        name: 'hensiktCheck',
-        message: EYupErrorMessage.PAAKREVD,
-        test: function (hensikt) {
-          const { parent } = this
-          if (parent.status === EKravStatus.AKTIV) {
-            return hensikt ? true : false
-          }
-          return true
-        },
-      }),
-      versjonEndringer: yup.string().test({
-        name: 'versjonEndringerCheck',
-        message: EYupErrorMessage.PAAKREVD,
-        test: function (versjonEndringer) {
-          const { parent } = this
-          if (parent.status === EKravStatus.AKTIV) {
-            if (!newKrav && krav && krav.kravVersjon > 1) {
-              return versjonEndringer ? true : false
-            }
-          }
-          return true
-        },
-      }),
-      regelverk: yup.array().test({
-        name: 'regelverkCheck',
-        message: EYupErrorMessage.PAAKREVD,
-        test: function (regelverk) {
-          const { parent } = this
-          if (parent.status === EKravStatus.AKTIV) {
-            return regelverk && regelverk.length > 0 ? true : false
-          }
-          return true
-        },
-      }),
-      varslingsadresser: yup.array().test({
-        name: 'varslingsadresserCheck',
-        message: EYupErrorMessage.PAAKREVD,
-        test: function (varslingsadresser) {
-          const { parent } = this
-          if (parent.status === EKravStatus.AKTIV) {
-            return varslingsadresser && varslingsadresser.length > 0 ? true : false
-          }
-          return true
-        },
-      }),
-      status: yup.string().test({
-        name: 'statusCheck',
-        message:
-          'Det er ikke lov å sette versjonen til utgått. Det eksistere en aktiv versjon som er lavere enn denne versjonen',
-        test: function (status) {
-          const { parent } = this
-          const nyesteAktivKravVersjon = alleKravVersjoner.filter(
-            (k) => k.kravStatus === EKravStatus.AKTIV
-          )
-          if (
-            status === EKravStatus.UTGAATT &&
-            nyesteAktivKravVersjon.length >= 1 &&
-            parent.kravVersjon > nyesteAktivKravVersjon[0].kravVersjon
-          ) {
-            return false
-          }
-          return true
-        },
-      }),
-    })
 
   const submit = async (krav: TKravQL) => {
     console.log('sumbit')
@@ -219,7 +127,10 @@ export const KravEditPage = () => {
 
   useEffect(() => {
     // hent krav på ny ved avbryt ny versjon
-    if (!edit && !krav?.id && krav?.nyKravVersjon) reloadKrav
+    if (!edit && !krav?.id && krav?.nyKravVersjon) {
+      reloadKrav
+      navigate(`/krav/${krav.kravNummer}/${krav.kravVersjon}`)
+    }
   }, [edit])
 
   useEffect(() => {
@@ -277,7 +188,7 @@ export const KravEditPage = () => {
                   : kravMapToFormVal(krav as TKravQL)
               }
               onSubmit={submit}
-              validationSchema={kravSchema()}
+              validationSchema={kravSchema({ newKrav, krav, alleKravVersjoner })}
             >
               {({
                 values,
